Adds a new patch called "Delayed Pulse"
This patch creates a pulse that works like the usual Pulse macro, but also takes in a Delay Amount and delays the pulse by that amount.
Useful when you want to pulse multiple things on canvas with different intervals. e.g. The wave-y effect when you post on someone's wall on Paper app.